refactor(fmt): extract shared format_count/format_duration into a fgumi-fmt leaf crate
`format_count` was duplicated in fgumi-cli-common and fgumi-metrics (same output,
two implementations), and `format_duration` was duplicated byte-for-byte in
fgumi-cli-common and fgumi-bam-io. Those three crates sit at different layers and
do not depend on one another, so there was no shared home for the helpers.

Add a dependency-free `fgumi-fmt` leaf crate with one canonical `format_count`
and `format_duration` (plus rstest tables + doctests), and have each consumer
re-export or import from it:

  - fgumi-cli-common: `pub use fgumi_fmt::{format_count, format_duration}` (keeps
    `fgumi_cli_common::format_count`/`format_duration` and `format_rate` working).
  - fgumi-metrics: `pub use fgumi_fmt::format_count` in `rejection` (keeps
    `fgumi_metrics::rejection::format_count` and `fgumi_metrics::format_count`).
  - fgumi-bam-io: `use fgumi_fmt::format_duration` in `progress` (drops the
    private copy).

No behavior change; all existing call paths keep resolving via the re-exports.
