+++ crates/fgumi-cli-common/Cargo.toml
@@ -12,6 +12,7 @@ anyhow = "1.0.102"
 bytesize = "2.3"
 clap = { version = "4", features = ["derive", "string"] }
 enum_dispatch = "0.3.13"
+fgumi-fmt = { workspace = true }
 log = "0.4"
 num_cpus = "1.16"
 sysinfo = { version = "0.38", default-features = false, features = ["system"] }
