+++ crates/fgumi-metrics/Cargo.toml
@@ -11,6 +11,7 @@ license.workspace = true
 serde = { version = "1.0", features = ["derive"] }
 anyhow = "1.0"
 fgoxide = "0.6.0"
+fgumi-fmt = { workspace = true }
 noodles = { version = "0.111.0", features = ["sam"], optional = true }
 fgumi-raw-bam = { workspace = true, optional = true }
